+++ source/Api/ArchivedMessages/ArchivedMessageGetDocumentListener.cs
@@ -16,7 +16,8 @@ using System;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Energinet.DataHub.EDI.Application.ArchivedMessages;
+using ArchivedMessages.Interfaces;
+using Energinet.DataHub.EDI.ArchivedMessages.Application;
 using MediatR;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
