+++ source/ArchivedMessages.Infrastructure/ArchivedMessageEntityConfiguration.cs
@@ -14,7 +14,7 @@
 
 using System;
 using System.IO;
-using Energinet.DataHub.EDI.Domain.ArchivedMessages;
+using Energinet.DataHub.EDI.ArchivedMessages.Application;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
