+++ source/ArchivedMessages.Infrastructure/Configuration/DataAccess/ArchivedMessagesContext.cs
@@ -1,0 +1,43 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using Energinet.DataHub.EDI.ArchivedMessages.Application;
+using Energinet.DataHub.EDI.Infrastructure.ArchivedMessages;
+using Microsoft.EntityFrameworkCore;
+
+namespace Energinet.DataHub.EDI.ArchivedMessages.Infrastructure.Configuration.DataAccess
+{
+    public class ArchivedMessagesContext : DbContext
+    {
+        #nullable disable
+        public ArchivedMessagesContext(DbContextOptions<ArchivedMessagesContext> options)
+            : base(options)
+        {
+        }
+
+        public ArchivedMessagesContext()
+        {
+        }
+
+        public DbSet<ArchivedMessage> ArchivedMessages { get; private set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            if (modelBuilder == null) throw new ArgumentNullException(nameof(modelBuilder));
+
+            modelBuilder.ApplyConfiguration(new ArchivedMessageEntityConfiguration());
+        }
+    }
+}
