+++ source/Infrastructure/Configuration/CompositionRoot.cs
@@ -20,6 +20,7 @@ using Energinet.DataHub.EDI.Application.Actors;
 using Energinet.DataHub.EDI.Application.Configuration;
 using Energinet.DataHub.EDI.Application.Configuration.Authentication;
 using Energinet.DataHub.EDI.Application.GridAreas;
+using Energinet.DataHub.EDI.ArchivedMessages.Infrastructure;
 using Energinet.DataHub.EDI.Common;
 using Energinet.DataHub.EDI.Common.Configuration;
 using Energinet.DataHub.EDI.Common.DataAccess;
