+++ source/Infrastructure/Configuration/DataAccess/B2BContext.cs
@@ -14,10 +14,8 @@
 
 using System;
 using Energinet.DataHub.EDI.Common.Actors;
-using Energinet.DataHub.EDI.Domain.ArchivedMessages;
 using Energinet.DataHub.EDI.Domain.GridAreas;
 using Energinet.DataHub.EDI.Infrastructure.Actors;
-using Energinet.DataHub.EDI.Infrastructure.ArchivedMessages;
 using Energinet.DataHub.EDI.Infrastructure.GridAreas;
 using Energinet.DataHub.EDI.Infrastructure.InboxEvents;
 using Energinet.DataHub.EDI.Infrastructure.IncomingMessages;
@@ -37,8 +35,6 @@ namespace Energinet.DataHub.EDI.Infrastructure.Configuration.DataAccess
         {
         }
 
-        public DbSet<ArchivedMessage> ArchivedMessages { get; private set; }
-
         public DbSet<ReceivedInboxEvent> ReceivedInboxEvents { get; private set; }
 
         public DbSet<TransactionIdForSender> TransactionIds { get; private set; }
@@ -53,7 +49,6 @@ namespace Energinet.DataHub.EDI.Infrastructure.Configuration.DataAccess
         {
             if (modelBuilder == null) throw new ArgumentNullException(nameof(modelBuilder));
 
-            modelBuilder.ApplyConfiguration(new ArchivedMessageEntityConfiguration());
             modelBuilder.ApplyConfiguration(new ReceivedInboxEventEntityConfiguration());
             modelBuilder.ApplyConfiguration(new TransactionIdEntityConfiguration());
             modelBuilder.ApplyConfiguration(new MessageIdEntityConfiguration());
