+++ source/Infrastructure/Infrastructure.csproj
@@ -58,6 +58,7 @@ limitations under the License.
   </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\ArchivedMessages.Infrastructure\ArchivedMessages.Infrastructure.csproj" />
       <ProjectReference Include="..\OutgoingMessages.Infrastructure\OutgoingMessages.Infrastructure.csproj" />
       <ProjectReference Include="..\Infrastructure.DocumentValidation\Infrastructure.DocumentValidation.csproj" />
       <ProjectReference Include="..\Infrastructure.CimMessageAdapter\Infrastructure.CimMessageAdapter.csproj" />
@@ -66,6 +67,7 @@ limitations under the License.
       <ProjectReference Include="..\Process.Interfaces\Process.Interfaces.csproj" />
       <ProjectReference Include="..\RequestResponse\source\RequestResponse\RequestResponse.csproj" />
     </ItemGroup>
+
