+++ source/OutgoingMessages.Application/OutgoingMessages/PeekHandler.cs
@@ -16,7 +16,6 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Energinet.DataHub.EDI.Common;
-using Energinet.DataHub.EDI.Domain.ArchivedMessages;
 using Energinet.DataHub.EDI.OutgoingMessages.Contracts;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.OutgoingMessages;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.OutgoingMessages.Queueing;
@@ -34,22 +33,20 @@ public class PeekHandler : IRequestHandler<PeekCommand, PeekResult>
     private readonly DocumentFactory _documentFactory;
     private readonly IOutgoingMessageRepository _outgoingMessageRepository;
     private readonly ActorMessageQueueContext _actorMessageQueueContext;
-    private readonly IArchivedMessageRepository _archivedMessageRepository;
+    //private readonly IArchivedMessageRepository _archivedMessageRepository;
 
     public PeekHandler(
         IActorMessageQueueRepository actorMessageQueueRepository,
         IMarketDocumentRepository marketDocumentRepository,
         DocumentFactory documentFactory,
         IOutgoingMessageRepository outgoingMessageRepository,
-        ActorMessageQueueContext actorMessageQueueContext,
-        IArchivedMessageRepository archivedMessageRepository)
+        ActorMessageQueueContext actorMessageQueueContext)
     {
         _actorMessageQueueRepository = actorMessageQueueRepository;
         _marketDocumentRepository = marketDocumentRepository;
         _documentFactory = documentFactory;
         _outgoingMessageRepository = outgoingMessageRepository;
         _actorMessageQueueContext = actorMessageQueueContext;
-        _archivedMessageRepository = archivedMessageRepository;
     }
 
     public async Task<PeekResult> Handle(PeekCommand request, CancellationToken cancellationToken)
@@ -79,15 +76,15 @@ public class PeekHandler : IRequestHandler<PeekCommand, PeekResult>
             document = new MarketDocument(result, peekResult.BundleId);
             await _marketDocumentRepository.AddAsync(document).ConfigureAwait(false);
 
-            _archivedMessageRepository.Add(new ArchivedMessage(
-                peekResult.BundleId.Id.ToString(),
-                peekResult.BundleId.Id.ToString(),
-                outgoingMessageBundle.DocumentType.ToString(),
-                outgoingMessageBundle.SenderId.Value,
-                outgoingMessageBundle.Receiver.Number.Value,
-                timestamp,
-                outgoingMessageBundle.BusinessReason,
-                result));
+            // _archivedMessageRepository.Add(new ArchivedMessage(
+            //     peekResult.BundleId.Id.ToString(),
+            //     peekResult.BundleId.Id.ToString(),
+            //     outgoingMessageBundle.DocumentType.ToString(),
+            //     outgoingMessageBundle.SenderId.Value,
+            //     outgoingMessageBundle.Receiver.Number.Value,
+            //     timestamp,
+            //     outgoingMessageBundle.BusinessReason,
+            //     result));
         }
 
         return new PeekResult(document.Payload, document.BundleId.Id);
